Create contentsecuritypolicy.php
Andy, my name is tom jansto and in the process of looking into CSP for a joomla 3.x install i run i cam across this plugin.  i downloaded it and used it( and found a minor bug with it.) no matter which CSP policy directive i specified they all used the 'default-src' policy tag.  the uploaded file simply corrects them to correspond to the matching policy directives.

thanks for the project plugin, it simplifies and standardizes what i needed.
tom